+++ click_extra/colorize.py
@@ -326,9 +326,6 @@ class HelpOption(ExtraOption):
         """
         if value and not ctx.resilient_parsing:
             echo(ctx.get_help(), color=ctx.color)
-            # Do not just ctx.exit() as it will prevent callbacks defined on options
-            # to be called.
-            ctx.close()
             ctx.exit()