+++ click_extra/commands.py
@@ -24,6 +24,7 @@ from __future__ import annotations
 
 import logging
 from typing import Any, cast
+from typing_extensions import NoReturn
 
 import click
 import cloup
@@ -98,6 +99,18 @@ class ExtraContext(cloup.Context):
     def color(self) -> None:
         """Reset the color value so it defaults to inheritance from parent's."""
         self._color = None
+
+    def exit(self, code: int = 0) -> NoReturn:
+        """Exits the application with a given exit code.
+
+        Forces the context to close before exiting, so callbacks attached to parameters
+        will be called to clean up their state.
+
+        .. todo::
+            Propose this fix upstream to Click.
+        """
+        self.close()
+        return super().exit(code)
 
 
 def default_extra_params():