+++ click_extra/parameters.py
@@ -711,7 +711,4 @@ class ShowParamsOption(ExtraOption, ParamStructure):
         )
         echo(output, color=ctx.color)
 
-        # Do not just ctx.exit() as it will prevent callbacks defined on options
-        # to be called.
-        ctx.close()
         ctx.exit()